Add association results filtering.
This adds result filtering so that an association only returns instances that match the result class specified during the query. Such an example is the following:

HostedService needs to return VSMS and RPCS instances when queried using CIM_ManagedElement as the result class.  However, if CIM_VSMS is specified as the result class, then only the VSMS instance needs to be returned.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>

 std_association.c |   61 +++++++++++++++++++++++++++++++++++++++++++++++++++++-
 1 file changed, 60 insertions(+), 1 deletion(-)