Make the EO parser not depend on root/ibmsd namespace
This EO parser needs some serious work.  It's not even remotely thread-safe
and it's pretty crusty.  This doesn't make it any less thread safe than
it already was, and it makes us not break things that depend on EO parsing.
I'll take a TODO to fix this up a bit sometime in the future.

Signed-off-by: Dan Smith <danms@us.ibm.com>