Make configure check for broken sblim-cmpi-devel version
The sblim version of cmpift.h is broken.  We can detect this by checking
that the size of the structure on 64-bit platforms meets the CMPI standard
of 160 bytes.  When sblim fixes their version, this test will start to pass.

Testing of this would be appreciated, especially if someone with a 32-bit
machine could give it a whirl.

Signed-off-by: Dan Smith <danms@us.ibm.com>