Removed error messages that cause wrong error returns
The case that no handler was found for a request caused an
error as return. This is wrong as the not-found handler
tells only, that the provider is not responsible for the
given request. In scenarios with a general request causing
the call of several association providers that return
valid instances, the former behavior causes a wrong
result, as the CIMOM would return the error and revoke all
valid results.
Signed-off-by: Heidi Eckhart <heidieck@linux.vnet.ibm.com>

 std_association.c |    7 +------
 1 file changed, 1 insertion(+), 6 deletions(-)