[CU] Merge do_ref() and do_assoc()
The two functions do_ref() and do_assoc() run through
the same logic. The only difference are the returned
instances. By configuring the do_assoc() function with
an additional paramter (ref_rslt) to inidcate the type
of the returned instance(s) - either reference(names) or
associator(names) - the two functions can be merged.
Signed-off-by: Heidi Eckhart <heidieck@linux.vnet.ibm.com>

 std_association.c |  167 +++++++++++++++++++++++++++++-------------------------
 1 file changed, 91 insertions(+), 76 deletions(-)