[CU] Add --print-ind switch to indication_tester
Sometimes just knowing you got the indication isn't informative enough, especially for early testing, so this patch adds an option to have the indication text printed out.  Adding the flag to the server feels a bit wrong, but it is *far* cleaner than extending the appropriate __init__ functions, so I'm okay with it.

Signed-off-by: Jay Gagnon <grendel@linux.vnet.ibm.com>

 indication_tester.py |   11 +++++++++--
 1 file changed, 9 insertions(+), 2 deletions(-)