Fix indication_tester's filter deletion issue
A while ago we discovered that giving the same name to various tests in indication_tester doesn't work, even when the first test's filter, handler, and subscription should have been deleted once it was done running.  It turns out that indication_tester wasn't deleting the filter, due to some bad xml.  This fixes the xml and eliminates the need to use the -n switch unless you want multiple simultaneous tests.

Signed-off-by: Jay Gagnon <grendel@linux.vnet.ibm.com>

 indication_tester.py |    4 ++--
 1 file changed, 2 insertions(+), 2 deletions(-)