[CU] indication_tester.py enhancements
Add support for a port option.  If nothing is supplied, port 8000 is used.  Otherwise, the value supplied is added to 8000.  This allows for multiple copies of the indication_tester to run at once.

Update delete_inst_xml() to take an inst_name argument. The classname needs to be CIM_IndicationHandlerCIMXML, not CIM_IndicationHandler.  However, the instance is created with the instance name "%s%s" % (self.name, "Handler")

This has been tested with both Pegasus and sfcb.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>