Work around for sfcb's lack of association handling
If a association provider has multiple classes listed, and a user queries with
the super class for the classes listed in the registration, the provider is
not called each time for each class.  Instead, the provider is called once.

This logic detects whether the association classname specified for by the user
is an exact match of one of the classnames in the registration.  If it is,
the provider is only called once.  If the classname is a super class, then the
provider is called once for each class listed in the registration.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>